Slim down the css and start putting in some content
First, putting in the intro content was a pain.  But, a bigger problem
was duplication of the code that's now under <section>.  A new problem
to deal with is that <nav> is now a child of <section>, when a better
organization would be to have <nav> outside of <section>, and contain
the .toptitle (my name).